Modified GET endpoint to add specify rest endpoints for each cli.
When creating REST endpoints, the endpoint is constructed based on:
<dockerImageName>/<cliname>/<run or xmlspec>. However due to unacceptable
characters in the docker image name, the docker image name used in
the actual endpoint differs.To allow for automated querying and calling of
the endpoints the get endpoint will specify each available endpoint for each
image. This also allows for the same docker image to have multiple
versions without conflict in histomicstk.