[WIP] Docker image Caching (#172)
* Began to refactor Docker Image caching to use girder models

Upon startup HistomicsTk will generate rest enpoints for image clis
using data stored in the mongo database.
The restendpoints get, and delete allow adding or deleting docker images
from the girder mongo database. If an image is being added but doesnt exist
a docker pull is attempted. If an image is being deleted from the mongo database
there is an option to delete the image off the local machine. Manually deleting
an image via docker rmi, results in the the plugin deleting the image data
from the mongo data base upon the next get request.For each succesful loaded image
rest endpoints are generated for each cli.

* added comments, fixed cli handler names and python styling

* Added unit tests to verify caching of docker image data to the Mongo Database

Unit tests verify that the PUT and DELETE endpoints are capable of
pulling a docker image from docker hub and deleting a docker image
from the local machine.

* Modified GET endpoint to add specify rest endpoints for each cli.

When creating REST endpoints, the endpoint is constructed based on:
<dockerImageName>/<cliname>/<run or xmlspec>. However due to unacceptable
characters in the docker image name, the docker image name used in
the actual endpoint differs.To allow for automated querying and calling of
the endpoints the get endpoint will specify each available endpoint for each
image. This also allows for the same docker image to have multiple
versions without conflict in histomicstk.

* Added a unit test to check xmlspec rest endpoints created during runtime

* Addressed python styling and added a test

A test was added to verify graceful termination of the girder job
when an arbitrary docker image is added.