Once we have built and installed openslide, purge the pip cache and explicitly uninstall Pillow if it is installed.
Ask pip for the location of the pip cache and remove that directory.  I feel this is brittle and heavy-handed.

If Pillow is installed, it will look for the old jpeg libraries rather than the ones we have just built.  If the cache is not cleared, pip will just reinstall Pillow from the cache and still look for the wrong libraries.