Add a script to deploy via docker containers.
Two docker containers are built, one for girder_worker and one for HistomicsTK.  Scripts optionally set the hosts file to refer to the docker host for mongo and rabbitmq, if desired.  The docker containers are based on the Girder-2 branches of all relevant repositories.

The girder/dev roles have been reapportioned to girder/provision roles with changes to the database all in the provision role.

The build tests won't work in the multiple docker-container deployment, as those tests expect direct control of mongo.

Ansible scripts have been modified to use the current recommended parameters (e.g., become rather than sudo).

More testing is needed: the scripts work when deployed via vagrant and via docker when locally built and using default options.

The docker images need to be published so that they do not need to be built locally.