Add cdog - a constrained difference-of-gaussian filter (#288)
This commit implements a faster alternative to the constrained laplacian-of-gaussian filter (used for nuclei detection) based on a difference-of-guassians.

The implementation is similar in concepts to the SIFT interest point detection (See here). It divides the scale space into octaves and each octave is divided into a set of discrete levels such that the sigma/scale is doubled after each octave. The implementation also uses cascaded gaussian convolutions (G(a) * G(b) = G(sqrt(a^2 + b^2))) and downsamples the images by 2 at the end of each octave to keep the convolutional kernels small.

Commit history:

* Modify clog to return sigma values associated with maximal response

* Add implementation of cdog

* Fix style errors

* Add unit test sha files for clog and cdog

Also add an assert to check number of nuclei in nuclei_segmentation_test

* Add blob detection test to plugin.cmake

* Minor fix in nuclei_segmentation_test

* Commentout references to test files already referenced in another tests

* Fix a bug in nuclei_segmentation_test

* Fixed a bug in blob_detection_filters_test

* Fix style errors

* Mock skimage.transform in conf.py