Deploy pinned versions unless asked to use latest.
Also, deploy a CLI docker image by default.  The program that does this can also run tests on the CLI to ensure that the server, girder worker, and libraries are deployed properly.  Add --cli-test to the start or provision commands to run the CLI tests.  Add --no-cli to not add a CLI.  The CLI docker image version is either a pinned version or latest, using the same logic as the main docker images.