Simplify the docker tests.
The docker endpoints are used from the slicer_cli_web plugin, and are tested there.  These tests were duplicates, and didn't actually cover any local code.  The tests have been simplified to just check that the expected endpoints are present.

This also removes docker-py from the development requirements, as it was only used in the tests.