Work around user and group ID issues in Docker.
This merges the two set_hosts.py scripts into a new set_environment.py script that works for both the worker and girder roles.  This script sets up some /etc/hosts values (as before), can set the worker tmp_root if a command line flag is specified, can adjust the specified user's UID and GID, and adds the specified user to the host's docker group.  If the UID or GID of the user changes, all files are changed appropriately and the ~/.ansible directory is removed (somehow it remained stuck to the old UID).

The docker scripts have been modified to run the appropriate commands with the new groups, so the docker containers no longer need to be restarted to pick up the group changes.

This has been tested with a system with different UID, GID, and docker GID.  It would probably be good to test it with a system with a conflicting UID or GID.