Recover after auto-save error.
If auto save fails for any reason, it will never work again.  For instance, if a second tab is opened with the same image, and the auto-saved annotation is deleted, the annotation can no longer be auto-saved.  In this case, the save is immediately retired.

If there is an error for any other reason, the save is not tried again until something else (changing the annotations) triggers it.  For instance, this recovers if the server is briefly unreachable when the save is being attempted.