Fix race condition when rapidly changing active annotations
Because the active annotation was set *after* the model fetch response
resolved, it was possible to click on another annotation before response
completed.  In this case, the UI could result in an inconsistent state.
This change sets the active annotation synchronously (before the fetch).
If another click happens before the fetch returns, the original callback
is ignored.