Reduce the number of panel rerenders on mousemove
This optimizes the draw and annotation panels so that they don't
completely rerender when responding to a mousemove event.  The highlight
properties are still stored inside the annotation models to maintain
state in case a rerender *does* occur.  This should help with the
performance issues for fast firing mouse events.