Fix the active annotation when the event stream reconnects.
When the event stream reconnects, we check to see if there are any
changes to annotations.  If there is an active annotation, reattach to
it by the reloaded annotation model.  Prior to this, the active
annotation was no longer bound to the appropriate events, resulting in
not auto-saving.

Also, guard against undefined models in mouse highlighting that can
occur in this instance.

Remove the setup.cfg file and adjust the plugin.cmake file to track
changes in Girder.  Flake8 has an issue that it uses the wrong
setup.cfg.  I don't think we should be specifying a universal with with
the cython extensions, and we weren't actually using bumpversion.