Change how the command yaml is structured.
When a command in a yaml >- block, it is quoted differently than when it
is on the same line as the command: keyword.  It works with proper
quoting on one line, but fails for reasons I don't fully get in multiple
lines.  Switch to the one-line version to get the expected quoting
behavior.

Also, do not log ansible commands where a password is used.