Refactor debouncing logic in annotated images dialog
The prior implementation debounced queries strictly by time.  This
implementation queues changes to the filter query and executes the
rest request after the server has responded to the last request.  This
change also prevents triggering a rest request when the query has not
changed by storing query state in a backbone model and triggering
updates on change events on that model.