Fix race conditions in analysis client test
There were three race conditions occuring.

1. Values where injected into the control panel by the ImageView before
the control panels were rendered because `h:analysis` was being
triggered before an asynchronous render call.  This connects to the new
event slicer_cli_web that is triggered inside the render function
itself.

2. The output file test was not correctly waiting for the input element
to change.

3. Setting the ROI sometimes occurs before the image feature layer is
created.  This protects against that and make the draw a no-op because
the region is always drawn when the feature layer is generated.

Depends on https://github.com/girder/slicer_cli_web/pull/58