Add some more options for how dask does multiprocessing.
The scheduler_address can now a special string to use non-distributed
multiprocessing.

If 'thread' (or 'threaded' or 'threads'), a ThreadPool is used with
num_threads_per_worker as the number of threads.  Values <= 0 use the
number of physical cores plus that number.

If 'multi' (or 'multiprocessing' or 'process"), a Python multiprocessing
pool is used with num_workers as the number of processes.  Values <= 0
use the number of physical cores plus that number.

If the scheduler_address is blank or an ip address, the current behavior
is maintained.  This now allows '0' as num_workers to use the number of
phyiscal cores, and values < 1 for num_threads_per_worker to let dask
compute how many to use.