Add a permissions editor dialog for annotations
This the changes from https://github.com/girder/large_image/pull/282
to add a "Permissions" button to the edit annotation dialog for users
with admin access to the annotation document.  I added the button to the
annotation editor dialog to avoid cluttering the main panel.  I know
this isn't easiest place to find, but I couldn't come up with anything
better.