Remove colinear points from annotations.
When we detect nuclei and generate a polygon of the perimeter, this
removes colinear points that do not alter the shape of the polygon.  In
a simple test of around 325,000 nuclei, this reduces the total number of
vertices from 24,094,710 to 13,954,237.

This also discarded around 1,400 degenerate polygons that had no area.
The bulk of these consisted of three points; for example: [18437.0,
11894.0, 0.0], [18438.0, 11895.0, 0.0], [18437.0, 11894.0, 0.0].