Improve the ability to run tests inside the docker container.
Use iptables to allow girder tests to reach the mongo container as if it
were on localhost from inside the main docker container.

Install some extra utilities, such as npx and eslint in the main docker
to allow testing.

With this change, after starting the docker group with
./deploy_docker.py, the following should result in all HistomicsTK tests
passing:

```
docker exec -it histomicstk_histomicstk bash -c "cd ../build && cmake
-DRUN_CORE_TESTS:BOOL=OFF -DTEST_PLUGINS:STRING=HistomicsTK ../girder &&
make -j && ctest --output-on-failure"
```