Attempt to prevent the context menu from overflowing
This adjusts the y coordinate of the context menu so that the bottom is
no lower than the bottom edge of the visible window.  The context menu
can still overflow if the vertical size of the window is too small to
fit the context menu.  If this becomes an issue we can make further
adjustments, e.g. put it in a scrolling container or add a "show more"
button.