Only trigger a mouse click action on a single element
Prior to this change, the events were triggered on *all* elements under
the mouse cursor.  In the event of a right click, the last element that
happened to be triggered by geojs was the one that was eventually
highlighted.  For multi-select, all elements were being toggled
resulting in a confusing behavior.

Now only a single element is chosen based on a fast heuristic--the
polygon with a vertex nearest the point clicked.  We can replace the
heuristic easily, but this computation necessarily adds processing that
must occur during the click event, so it will slow down the context menu
action (perhaps un-noticeably so).