Keep "show all annotation" state when opening new images
This is the simplest possible implementation.  It stores the state as a
private variable on the annotation selector view.  When any display
state is manually toggled, it is switched off.

This isn't tested because at the moment it would require a lot of code
(and testing time) to open a new image to check what is effectively one
additional code branch.