Don't show annotations that are too slow to load.
If you ask to show an annotation on a slide that is slow to load, then
switch to another slide, the annotation would appear on the now wrong
slide.  Similarly, if you show an annotation that is slow to load, then
hide all annotations, that annotation could still appear once it was
loaded.

This checks to make sure it is still appropriate to show an annotation
when it finishes loading.