Fix merge #599 expose tissue seg mag reinhard stats (#600)
All though CI build of PR #599 succeeded, it oddly fails after merging to master. This PR tries to fix that. We identified that the reason for this might be due to the change in the default value of the `resample` parameter of [large_image.TilseSource.tileIterator](https://github.com/girder/large_image/pull/336) to True.

Commit history:
---------------------
* Revert "Expose tissue_seg_mag parameter of sample_pixels in reinhard_stats"
* Explicitly specify key-value pairs of all opt params
* Update groundtruth reinhard mean
* Update reinhard sigma groundtruth