Reduce files in the HistomicsTK docker image.
When we added some code to make it easier to run cmake tests in the
HistomicsTK docker image, this also added ~40,000 files to the image.
Because of how docker works, we could possibly chown these files on
initial startup, and this process can be slow on some systems.

This undoes those additions -- some files are chown to root; others are
removed.

This means, to run the full tests, you will need to run `girder-install
web --dev` at least once within the docker.  The `devops/build.sh`
script has been updated to reflect this.