Fix transparency slider issue #618.
This issue was caused by only allowing changes to opacity for annotation
sets with less than 1000 elements, but allowing editing with annotations
with up to 5000 elements.

We should test changing opacity on very large sets, as it probably is
still performant and the limit could be raised further or removed
entirely.

For now, the limit has been raised to be the same as the editable limit.