Speed up draw widget rendering by using native tooltips.
When we render the draw widget with the list of annotation elements,
about 3/4 of the render time is spent setting up Bootstrap tooltips.
While these look nicer than native tooltips, it isn't worth multiple
seconds of delay.

We could have a threshold where if there are a few elements, we use
fancy tooltips, and don't for many elements.  Here, I've just
de-bootstrapped the tooltips.