Restrict some endpoints to require a logged in user.
This adds a girder configuration value to control if the restriction is
added.  When turned on via the grider config value `[histomicstk]`
`restrict_downloads: True`, all standard girder download endpoints and
the large_image `item/{id}/tiles/images/{image}` endpoint are restricted
to logged-in-users.  This is in addition to ordinary ACL restrictions.

The restriction is turned on for our standard deployment.