Only show innermost polygon in annotation popover
This uses the same heuristic as the context menu to determine the
innermost element--the element contain a vertex closest to the mouse
position.  The update is throttled to 100ms/call to prevent blocking too
much on mouse move.  Having many polygons in a single spot will likely
cause performance issues, but hopefully that is unlikely to occur in
real world scenarios.

Fixes #632