Add some code to help build wheels.
- `Dockerfile-wheels` builds the wheels for linux.
- `build_wheels.sh` builds the docker file and copies the wheels to a
  temp wheels directory.
- `test_wheels.py` does a minimal check to ensure that the wheels work.

This also uses the published version of ctk-cli, patching the one class
that we had changed.  This avoids either havign to build a wheel for
ctk-cli or require an old version of pip to install from a git
repo.

As a future step, we should use setuptools-scm and integrate this into
CI.