Make sure annotations are saved.
We have a bug when saving annotations to a slow database or an
intermittently connected server.  Specifically, when we are saving an
annotation and another change occurs, we were not saving again (to
prevent race conditions).  However, this means the second save might
never happen.  Similarly, if the save failed, we would not retry it.

This fixes the issue by marking that we need to call save again after
the first save request completes (success or fail).  In the event of
failure, retry the save after a delay.