Refactor the positive pixel count example.
This now can handle arbitrary polygon regions and output images of any
size.  It parallelizes using dask if the region is big enough.  The
default is to output an annotation with a tiled tiff.  For a different
format, specify a file with a different extension (this will be limited
in size based on that format).